Added mocks for VMs
