+++ service/discovery/azure/compute.go
