+++ service/discovery/azure/storage.go
