+++ service/discovery/azure/storage_test.go
@@ -1,0 +1,78 @@
+package azure_test
+
+import (
+	"net/http"
+	"testing"
+
+	"clouditor.io/clouditor/service/discovery/azure"
+	"clouditor.io/clouditor/voc"
+	"github.com/stretchr/testify/assert"
+)
+
+type mockStorageSender struct {
+	mockSender
+}
+
+func (m mockStorageSender) Do(req *http.Request) (res *http.Response, err error) {
+	var handled bool
+
+	if res, handled, err = m.doSubscriptions(req); handled {
+		return
+	}
+
+	if req.URL.Path == "/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Storage/storageAccounts" {
+		res, err = createResponse(map[string]interface{}{
+			"value": &[]map[string]interface{}{
+				{
+					"id":       "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Storage/storageAccounts/account1",
+					"name":     "account1",
+					"location": "eastus",
+					"properties": map[string]interface{}{
+						"creationTime": "2017-05-24T13:28:53.4540398Z",
+						"primaryEndpoints": map[string]interface{}{
+							"blob": "https://account1.blob.core.windows.net/",
+						},
+						"encryption": map[string]interface{}{
+							"services": map[string]interface{}{
+								"file": map[string]interface{}{
+									"keyType":         "Account",
+									"enabled":         true,
+									"lastEnabledTime": "2019-12-11T20:49:31.7036140Z",
+								},
+								"blob": map[string]interface{}{
+									"keyType":         "Account",
+									"enabled":         true,
+									"lastEnabledTime": "2019-12-11T20:49:31.7036140Z",
+								},
+							},
+							"keySource": "Microsoft.Storage",
+						},
+					},
+				},
+			},
+		}, 200)
+	} else {
+		res, err = createResponse(map[string]interface{}{}, 404)
+		log.Errorf("Not handling mock for %s yet", req.URL.Path)
+	}
+
+	return
+}
+
+func TestListStorage(t *testing.T) {
+	d := azure.NewAzureStorageDiscovery(
+		azure.WithSender(&mockStorageSender{}),
+		azure.WithAuthorizer(&mockAuthorizer{}),
+	)
+
+	list, err := d.List()
+
+	assert.Nil(t, err)
+	assert.NotNil(t, list)
+	assert.Equal(t, 1, len(list))
+
+	storage, ok := list[0].(*voc.ObjectStorageResource)
+
+	assert.True(t, ok)
+	assert.Equal(t, "account1", storage.Name)
+}
